backend/requirements.txt
# Framework Web
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Banco de Dados
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Autenticação
PyJWT==2.8.0
bcrypt==4.1.1

# Validação
email-validator==2.1.0

# Variáveis de Ambiente
python-dotenv==1.0.0

# Testes
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Lint e Formatação
black==23.12.0
flake8==6.1.0
pylint==3.0.3

# Testes E2E
selenium==4.16.0

# Utilitários
python-dateutil==2.8.2
